Made the User list fully functional
The settings.html is now splitted in 3 parts to populate the
table with the data from the settings file.

The add/remove buttons will now update the names of the input
field accordingly and also update the count of users in the form

The form now writes the users back to the settings file.